Calender_Agent/pyproject.toml
[project]
name = "newbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.177.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.5.4",
    "mcp>=1.12.1",
    "mcp-use>=1.3.7",
    "pytz>=2025.2",
]
